requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
pandas
s3fs # this is for interacting with data stored in s3.
smart-open
pillow
torch
torchvision
torchaudio
sentence-transformers
transformers
boto3 # this library allows you to connect to and use AWS.We use to connect to s3.
einops
supabase
psycopg2
python-dotenv
pyarrow
tqdm
scikit-learn
matplotlib
umap-learn
fastapi
uvicorn
pydantic
pytest
openai